Refactor: remove "_proto" from "image_proto.py"
